Further work on the effects modifiers. Implemented a light-weight effect class for storing, modifying, and passing around augment characteristics. Next up implementing the new framework thru-out the augments.
(cherry picked from commit fbf0abb3db5309aa4d0b17526efb71355aa59321)

+++ src/main/kotlin/me/fzzyhmstrs/amethyst_imbuement/entity/ModifiableDamageEntity.kt
@@ -2,5 +2,5 @@ package me.fzzyhmstrs.amethyst_imbuement.entity
 
 interface ModifiableDamageEntity {
 
-    var damageModifier: Int
+    var damageModifier: Float
 }
